move the code that tests various configurations to it's own class
